backend/pyproject.toml
[tool.poetry]
name = "rag-edu-generator"
version = "0.1.0"
description = "RAG-powered educational content generator backend"
authors = ["Developer"]
readme = "README.md"
packages = [{include = "fastapi_backend"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
llama-index = "^0.10.0"
llama-index-vector-stores-pinecone = "^0.1.0"
llama-index-embeddings-openai = "^0.1.0"
llama-index-llms-openai = "^0.1.0"
llama-cloud = "0.0.6"
pymupdf = "^1.23.0"
pinecone-client = "^3.0.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
numpy = "^1.24.0"
setuptools = "^80.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
ruff = "^0.1.6"

[tool.poetry.scripts]
backend = "fastapi_backend.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

